+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - **Feature:** Added `MiniAppMessageBridge.sendJsonToHostApp` interface to provide `Universal Bridge` for sending messages from a MiniApp to the HostApp.
 - **Feature:** Added `NativeEventType.MINIAPP_RECEIVE_JSON_INFO` to provide `Universal Bridge` for receiving messages from the HostApp to MiniApps.
 - **Feature:** Added `sendJsonToMiniApp` in `MiniAppView` and `MiniAppViewImpl` for receiving messages from the HostApp to MiniApps.
-- **Fix:** Prevent file creation with zero byte with appending a single dot(.) at the end of file's name while downloading a file using `MiniAppFileDownloader` in Android 10.
+- **Update:** Added `allEmailList` field in `Contact` to support multiple emails of a specific contact.
+- **Fix:** Prevent file creation with zero byte while downloading a file using `MiniAppFileDownloader` in the offline mode and also support for Android 10.
 
 **Sample App**
 - **Feature:** Settings is moved to it's own tab.